[feat] 상세 피드 / 댓글 생성

+++ app/src/main/java/com/android/go/sopt/winey/data/service/FeedService.kt
@@ -56,7 +56,7 @@ interface FeedService {
 
     @POST("comment/{feedId}")
     suspend fun postComment(
-        @Path("feedId") feedId: Long,
+        @Path("feedId") feedId: Int,
         @Body requestPostCommentDto: RequestPostCommentDto
     ): BaseResponse<ResponsePostCommentDto>
 }
